js/jsd/test/test_bug602003.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=507448
-->
<head>
  <title>Test for Bug 602003</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=507448">Mozilla Bug 507448</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 602003 **/

// This is somewhat unfortunate: jsd only deals with scripts that have a
// nonzero line number, so we can't just createElement a script here.
// So break the test up into three <script>s, of which the middle one has our test functions.

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var jsdIDebuggerService = Components.interfaces.jsdIDebuggerService;
var jsd = Components.classes['@mozilla.org/js/jsd/debugger-service;1']
                    .getService(jsdIDebuggerService);
var jsdOn = jsd.isOn;
if (!jsdOn) {
  jsd.on();
  ok(jsd.isOn, "JSD should be running.");
}
</script>
<script>
  function g(a,b) { return a + b }
</script>
<script>
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var script = jsd.wrapValue(g).script;

  // Test the script start/end PC APIs.
  var start = script.getFirstValidPC();
  var end = script.getEndValidPC();

  // Start PC should be 1 for a function because it starts with JSOP_BEGIN.
  is(start, 1, "Start PC should be 1");

  // End PC should be something greater than 1, and not huge. Changes
  // in the bytecode will change this, so we'll just be approximate.
  ok(1 < end && end < 100, "End PC doesn't seem sane.");

  if (!jsdOn) {
    jsd.off();
    ok(!jsd.isOn, "JSD shouldn't be running anymore.");
  }
</script>
</pre>
</body>
</html>
